+++ src/mocks.ts
@@ -2,6 +2,7 @@ import casual from 'casual';
 import { DMSP_BASE_URL, validateDmspId } from './resolvers/scalars/dmspId';
 import { validateOrcid } from './resolvers/scalars/orcid';
 import { ROR_DOMAIN, validateRor } from './resolvers/scalars/ror';
+import { mock as affiliationMock } from './mocks/affiliation';
 import { mock as userMock } from './mocks/user';
 import { mock as contributorRoleMock } from './mocks/contributorRole';
 
@@ -35,6 +36,7 @@ const scalarMocks = {
 
 export const mocks = {
   ...scalarMocks,
+  ...affiliationMock,
   ...contributorRoleMock,
   ...userMock
 };