+++ src/resolver.ts
@@ -5,6 +5,7 @@ import { dmspIdScalar } from './resolvers/scalars/dmspId';
 import { orcidScalar } from './resolvers/scalars/orcid';
 import { rorScalar } from './resolvers/scalars/ror';
 
+import { resolvers as affiliationResolvers } from './resolvers/affiliation';
 import { resolvers as contributorRoleResolvers } from './resolvers/contributorRole';
 import { resolvers as dmspResolvers } from './resolvers/dmsp';
 
@@ -16,6 +17,7 @@ const scalarResolvers = {
 
 export const resolvers: IResolvers = mergeResolvers([
   scalarResolvers,
+  affiliationResolvers,
   dmspResolvers,
   contributorRoleResolvers
 ]);