+++ src/types.ts
@@ -28,14 +28,57 @@ export type Scalars = {
 
 export type Affiliation = {
   __typename?: 'Affiliation';
-  affiliation_id?: Maybe<OrganizationIdentifier>;
+  acronyms?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
+  active: Scalars['Boolean']['output'];
+  addresses?: Maybe<Array<Maybe<AffiliationAddress>>>;
+  aliases?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
+  countryCode?: Maybe<Scalars['String']['output']>;
+  countryName?: Maybe<Scalars['String']['output']>;
+  displayName: Scalars['String']['output'];
+  domain?: Maybe<Scalars['String']['output']>;
+  funder: Scalars['Boolean']['output'];
+  id: Scalars['String']['output'];
+  links?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
+  name: Scalars['String']['output'];
+  relationships?: Maybe<Array<Maybe<AffiliationRelationship>>>;
+  types?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
+  wikipediaURL?: Maybe<Scalars['URL']['output']>;
+};
+
+export type AffiliationAddress = {
+  __typename?: 'AffiliationAddress';
+  city?: Maybe<Scalars['String']['output']>;
+  country?: Maybe<Scalars['String']['output']>;
+  countryGeonamesId?: Maybe<Scalars['Int']['output']>;
+  lat?: Maybe<Scalars['Float']['output']>;
+  lng?: Maybe<Scalars['Float']['output']>;
+  state?: Maybe<Scalars['String']['output']>;
+  stateCode?: Maybe<Scalars['String']['output']>;
+};
+
+export type AffiliationRelationship = {
+  __typename?: 'AffiliationRelationship';
+  id: Scalars['String']['output'];
+  name: Scalars['String']['output'];
+  type: Scalars['String']['output'];
+};
+
+export type AffiliationSearch = {
+  __typename?: 'AffiliationSearch';
+  aliases?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
+  countryCode?: Maybe<Scalars['String']['output']>;
+  countryName?: Maybe<Scalars['String']['output']>;
+  funder: Scalars['Boolean']['output'];
+  fundref?: Maybe<Scalars['String']['output']>;
+  id: Scalars['String']['output'];
+  links?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
   name: Scalars['String']['output'];
 };
 
 export type Contributor = Person & {
   __typename?: 'Contributor';
   contributorId?: Maybe<PersonIdentifier>;
-  dmproadmap_affiliation?: Maybe<Affiliation>;
+  dmproadmap_affiliation?: Maybe<DmpRoadmapAffiliation>;
   mbox?: Maybe<Scalars['String']['output']>;
   name: Scalars['String']['output'];
   role: Array<Scalars['String']['output']>;
@@ -71,6 +114,12 @@ export type ContributorRoleMutationResponse = {
   message: Scalars['String']['output'];
   /** Indicates whether the mutation was successful */
   success: Scalars['Boolean']['output'];
+};
+
+export type DmpRoadmapAffiliation = {
+  __typename?: 'DmpRoadmapAffiliation';
+  affiliation_id?: Maybe<OrganizationIdentifier>;
+  name: Scalars['String']['output'];
 };
 
 export type Dmsp = {
@@ -143,7 +192,7 @@ export type OrganizationIdentifier = {
 };
 
 export type Person = {
-  dmproadmap_affiliation?: Maybe<Affiliation>;
+  dmproadmap_affiliation?: Maybe<DmpRoadmapAffiliation>;
   mbox?: Maybe<Scalars['String']['output']>;
   name: Scalars['String']['output'];
 };
@@ -157,7 +206,7 @@ export type PersonIdentifier = {
 export type PrimaryContact = Person & {
   __typename?: 'PrimaryContact';
   contact_id?: Maybe<Identifier>;
-  dmproadmap_affiliation?: Maybe<Affiliation>;
+  dmproadmap_affiliation?: Maybe<DmpRoadmapAffiliation>;
   mbox?: Maybe<Scalars['String']['output']>;
   name: Scalars['String']['output'];
 };
@@ -165,6 +214,10 @@ export type PrimaryContact = Person & {
 export type Query = {
   __typename?: 'Query';
   _empty?: Maybe<Scalars['String']['output']>;
+  /** Retrieve a specific Affiliation by its ID */
+  affiliation?: Maybe<Affiliation>;
+  /** Perform a search for Affiliations matching the specified name */
+  affiliations?: Maybe<Array<Maybe<AffiliationSearch>>>;
   /** Get the contributor role by it's id */
   contributorRoleById?: Maybe<ContributorRole>;
   /** Get the contributor role by it's URL */
@@ -175,6 +228,17 @@ export type Query = {
   dmspById?: Maybe<SingleDmspResponse>;
   me?: Maybe<User>;
   users?: Maybe<Array<Maybe<User>>>;
+};
+
+
+export type QueryAffiliationArgs = {
+  affiliationId: Scalars['String']['input'];
+};
+
+
+export type QueryAffiliationsArgs = {
+  funderOnly?: InputMaybe<Scalars['Boolean']['input']>;
+  name: Scalars['String']['input'];
 };
 
 
@@ -312,15 +376,20 @@ export type ResolversInterfaceTypes<_RefType extends Record<string, unknown>> = 
 /** Mapping between all available schema types and the resolvers types */
 export type ResolversTypes = {
   Affiliation: ResolverTypeWrapper<Affiliation>;
+  AffiliationAddress: ResolverTypeWrapper<AffiliationAddress>;
+  AffiliationRelationship: ResolverTypeWrapper<AffiliationRelationship>;
+  AffiliationSearch: ResolverTypeWrapper<AffiliationSearch>;
   Boolean: ResolverTypeWrapper<Scalars['Boolean']['output']>;
   Contributor: ResolverTypeWrapper<Contributor>;
   ContributorRole: ResolverTypeWrapper<ContributorRoleModel>;
   ContributorRoleMutationResponse: ResolverTypeWrapper<Omit<ContributorRoleMutationResponse, 'contributorRole'> & { contributorRole?: Maybe<ResolversTypes['ContributorRole']> }>;
   DateTimeISO: ResolverTypeWrapper<Scalars['DateTimeISO']['output']>;
+  DmpRoadmapAffiliation: ResolverTypeWrapper<DmpRoadmapAffiliation>;
   Dmsp: ResolverTypeWrapper<DmspModel>;
   DmspId: ResolverTypeWrapper<Scalars['DmspId']['output']>;
   DmspIdentifier: ResolverTypeWrapper<DmspIdentifier>;
   EmailAddress: ResolverTypeWrapper<Scalars['EmailAddress']['output']>;
+  Float: ResolverTypeWrapper<Scalars['Float']['output']>;
   ID: ResolverTypeWrapper<Scalars['ID']['output']>;
   Identifier: ResolverTypeWrapper<Identifier>;
   Int: ResolverTypeWrapper<Scalars['Int']['output']>;
@@ -344,15 +413,20 @@ export type ResolversTypes = {
 /** Mapping between all available schema types and the resolvers parents */
 export type ResolversParentTypes = {
   Affiliation: Affiliation;
+  AffiliationAddress: AffiliationAddress;
+  AffiliationRelationship: AffiliationRelationship;
+  AffiliationSearch: AffiliationSearch;
   Boolean: Scalars['Boolean']['output'];
   Contributor: Contributor;
   ContributorRole: ContributorRoleModel;
   ContributorRoleMutationResponse: Omit<ContributorRoleMutationResponse, 'contributorRole'> & { contributorRole?: Maybe<ResolversParentTypes['ContributorRole']> };
   DateTimeISO: Scalars['DateTimeISO']['output'];
+  DmpRoadmapAffiliation: DmpRoadmapAffiliation;
   Dmsp: DmspModel;
   DmspId: Scalars['DmspId']['output'];
   DmspIdentifier: DmspIdentifier;
   EmailAddress: Scalars['EmailAddress']['output'];
+  Float: Scalars['Float']['output'];
   ID: Scalars['ID']['output'];
   Identifier: Identifier;
   Int: Scalars['Int']['output'];
@@ -372,14 +446,57 @@ export type ResolversParentTypes = {
 };
 
 export type AffiliationResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['Affiliation'] = ResolversParentTypes['Affiliation']> = {
-  affiliation_id?: Resolver<Maybe<ResolversTypes['OrganizationIdentifier']>, ParentType, ContextType>;
+  acronyms?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
+  active?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
+  addresses?: Resolver<Maybe<Array<Maybe<ResolversTypes['AffiliationAddress']>>>, ParentType, ContextType>;
+  aliases?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
+  countryCode?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  countryName?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  displayName?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  domain?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  funder?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
+  id?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  links?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
+  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  relationships?: Resolver<Maybe<Array<Maybe<ResolversTypes['AffiliationRelationship']>>>, ParentType, ContextType>;
+  types?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
+  wikipediaURL?: Resolver<Maybe<ResolversTypes['URL']>, ParentType, ContextType>;
+  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+};
+
+export type AffiliationAddressResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['AffiliationAddress'] = ResolversParentTypes['AffiliationAddress']> = {
+  city?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  country?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  countryGeonamesId?: Resolver<Maybe<ResolversTypes['Int']>, ParentType, ContextType>;
+  lat?: Resolver<Maybe<ResolversTypes['Float']>, ParentType, ContextType>;
+  lng?: Resolver<Maybe<ResolversTypes['Float']>, ParentType, ContextType>;
+  state?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  stateCode?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+};
+
+export type AffiliationRelationshipResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['AffiliationRelationship'] = ResolversParentTypes['AffiliationRelationship']> = {
+  id?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  type?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+};
+
+export type AffiliationSearchResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['AffiliationSearch'] = ResolversParentTypes['AffiliationSearch']> = {
+  aliases?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
+  countryCode?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  countryName?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  funder?: Resolver<ResolversTypes['Boolean'], ParentType, ContextType>;
+  fundref?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  id?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  links?: Resolver<Maybe<Array<Maybe<ResolversTypes['String']>>>, ParentType, ContextType>;
   name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
   __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
 };
 
 export type ContributorResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['Contributor'] = ResolversParentTypes['Contributor']> = {
   contributorId?: Resolver<Maybe<ResolversTypes['PersonIdentifier']>, ParentType, ContextType>;
-  dmproadmap_affiliation?: Resolver<Maybe<ResolversTypes['Affiliation']>, ParentType, ContextType>;
+  dmproadmap_affiliation?: Resolver<Maybe<ResolversTypes['DmpRoadmapAffiliation']>, ParentType, ContextType>;
   mbox?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
   role?: Resolver<Array<ResolversTypes['String']>, ParentType, ContextType>;
@@ -408,6 +525,12 @@ export type ContributorRoleMutationResponseResolvers<ContextType = MyContext, Pa
 export interface DateTimeIsoScalarConfig extends GraphQLScalarTypeConfig<ResolversTypes['DateTimeISO'], any> {
   name: 'DateTimeISO';
 }
+
+export type DmpRoadmapAffiliationResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['DmpRoadmapAffiliation'] = ResolversParentTypes['DmpRoadmapAffiliation']> = {
+  affiliation_id?: Resolver<Maybe<ResolversTypes['OrganizationIdentifier']>, ParentType, ContextType>;
+  name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
+  __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
+};
 
 export type DmspResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['Dmsp'] = ResolversParentTypes['Dmsp']> = {
   contact?: Resolver<ResolversTypes['PrimaryContact'], ParentType, ContextType>;
@@ -466,7 +589,7 @@ export type OrganizationIdentifierResolvers<ContextType = MyContext, ParentType 
 
 export type PersonResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['Person'] = ResolversParentTypes['Person']> = {
   __resolveType: TypeResolveFn<'Contributor' | 'PrimaryContact', ParentType, ContextType>;
-  dmproadmap_affiliation?: Resolver<Maybe<ResolversTypes['Affiliation']>, ParentType, ContextType>;
+  dmproadmap_affiliation?: Resolver<Maybe<ResolversTypes['DmpRoadmapAffiliation']>, ParentType, ContextType>;
   mbox?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
 };
@@ -479,7 +602,7 @@ export type PersonIdentifierResolvers<ContextType = MyContext, ParentType extend
 
 export type PrimaryContactResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['PrimaryContact'] = ResolversParentTypes['PrimaryContact']> = {
   contact_id?: Resolver<Maybe<ResolversTypes['Identifier']>, ParentType, ContextType>;
-  dmproadmap_affiliation?: Resolver<Maybe<ResolversTypes['Affiliation']>, ParentType, ContextType>;
+  dmproadmap_affiliation?: Resolver<Maybe<ResolversTypes['DmpRoadmapAffiliation']>, ParentType, ContextType>;
   mbox?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
   name?: Resolver<ResolversTypes['String'], ParentType, ContextType>;
   __isTypeOf?: IsTypeOfResolverFn<ParentType, ContextType>;
@@ -487,6 +610,8 @@ export type PrimaryContactResolvers<ContextType = MyContext, ParentType extends 
 
 export type QueryResolvers<ContextType = MyContext, ParentType extends ResolversParentTypes['Query'] = ResolversParentTypes['Query']> = {
   _empty?: Resolver<Maybe<ResolversTypes['String']>, ParentType, ContextType>;
+  affiliation?: Resolver<Maybe<ResolversTypes['Affiliation']>, ParentType, ContextType, RequireFields<QueryAffiliationArgs, 'affiliationId'>>;
+  affiliations?: Resolver<Maybe<Array<Maybe<ResolversTypes['AffiliationSearch']>>>, ParentType, ContextType, RequireFields<QueryAffiliationsArgs, 'name'>>;
   contributorRoleById?: Resolver<Maybe<ResolversTypes['ContributorRole']>, ParentType, ContextType, RequireFields<QueryContributorRoleByIdArgs, 'contributorRoleId'>>;
   contributorRoleByURL?: Resolver<Maybe<ResolversTypes['ContributorRole']>, ParentType, ContextType, RequireFields<QueryContributorRoleByUrlArgs, 'contributorRoleURL'>>;
   contributorRoles?: Resolver<Maybe<Array<Maybe<ResolversTypes['ContributorRole']>>>, ParentType, ContextType>;
@@ -533,10 +658,14 @@ export type UserResolvers<ContextType = MyContext, ParentType extends ResolversP
 
 export type Resolvers<ContextType = MyContext> = {
   Affiliation?: AffiliationResolvers<ContextType>;
+  AffiliationAddress?: AffiliationAddressResolvers<ContextType>;
+  AffiliationRelationship?: AffiliationRelationshipResolvers<ContextType>;
+  AffiliationSearch?: AffiliationSearchResolvers<ContextType>;
   Contributor?: ContributorResolvers<ContextType>;
   ContributorRole?: ContributorRoleResolvers<ContextType>;
   ContributorRoleMutationResponse?: ContributorRoleMutationResponseResolvers<ContextType>;
   DateTimeISO?: GraphQLScalarType;
+  DmpRoadmapAffiliation?: DmpRoadmapAffiliationResolvers<ContextType>;
   Dmsp?: DmspResolvers<ContextType>;
   DmspId?: GraphQLScalarType;
   DmspIdentifier?: DmspIdentifierResolvers<ContextType>;